+++ src/Http/Controllers/TokenAuthController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Log;
+use InternetGuru\LaravelCommon\Contracts\ReCaptchaInterface;
 
 class TokenAuthController extends Controller
 {
@@ -28,8 +29,12 @@ class TokenAuthController extends Controller
     /**
      * Send the token auth link to the user based on the form email input
      */
-    public function handleTokenAuthSendForm(Request $request): RedirectResponse
+    public function handleTokenAuthSendForm(Request $request, ReCaptchaInterface $recaptcha): RedirectResponse
     {
+        $recaptcha->validate($request);
+        $request->validate([
+            'email' => 'required|email|max:255',
+        ]);
         try {
             $user = User::where('email', $request->input('email'))->firstOrFail();